backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pillow==11.0.0
